Refactor snapshot tests to inherit common traits
